Rename logging module
