+++ src/reefcraft/utils/__init__.py
@@ -1,1 +1,5 @@
 """Utility helpers used throughout the Reefcraft project."""
+
+from .logger import configure_logging, logger
+
+__all__ = ["configure_logging", "logger"]
